requirements.txt
aioredis==1.2.0
amqp==2.3.2
asgiref==2.3.2
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.7.1
Automat==0.7.0
bcrypt==3.1.7
billiard==3.5.0.4
celery==4.2.0
certifi==2018.4.16
cffi==1.12.3
channels==2.1.3
channels-redis==2.3.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.7
daphne==2.2.5
defusedxml==0.5.0
Django==2.0.1
django-celery-beat==1.1.1
django-celery-results==1.0.1
django-ckeditor==5.4.0
django-cleanup==2.1.0
django-jet==1.0.8
django-js-asset==1.1.0
django-modeltranslation==0.13.1
django-rosetta==0.9.3
django-widget-tweaks==1.4.2
gunicorn==19.7.1
hiredis==1.0.0
hyperlink==19.0.0
idna==2.7
incremental==17.5.0
kombu==4.2.1
msgpack==0.5.6
oauthlib==2.1.0
Pillow==5.1.0
#pkg-resources==0.0.0
polib==1.1.0
psycopg2==2.7.4
psycopg2-binary==2.7.5
pycparser==2.19
PyHamcrest==1.9.0
PyJWT==1.6.4
python3-openid==3.1.0
pytz==2018.4
redis==2.10.6
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
Twisted==19.2.1
txaio==18.8.1
Unidecode==1.0.23
urllib3==1.23
uWSGI==2.0.17
vine==1.1.4
zope.interface==4.6.0
